refactor(auth): nest internal packages under pkg/auth
Move packages/auth/internal/* to packages/auth/pkg/auth/internal/* so the
implementation lives inside the consumer-facing pkg tree. Consumers keep
importing the stable facade at .../packages/auth/pkg/auth.

Go's internal visibility now scopes the implementation to pkg/auth alone
(pkg/types and pkg/tests can no longer reach it), and image builds that
COPY ./auth/pkg pick up the implementation for free — drop the separate
./auth/internal COPY that #3323 added to unbreak image builds after #3314.
