+++ packages/auth/pkg/auth/internal/service/service.go
@@ -11,9 +11,9 @@ import (
 	"github.com/google/uuid"
 	"github.com/redis/go-redis/v9"
 
-	"github.com/e2b-dev/infra/packages/auth/internal/authcontext"
-	internalauthteam "github.com/e2b-dev/infra/packages/auth/internal/team"
-	"github.com/e2b-dev/infra/packages/auth/internal/token"
+	"github.com/e2b-dev/infra/packages/auth/pkg/auth/internal/authcontext"
+	internalauthteam "github.com/e2b-dev/infra/packages/auth/pkg/auth/internal/team"
+	"github.com/e2b-dev/infra/packages/auth/pkg/auth/internal/token"
 	"github.com/e2b-dev/infra/packages/auth/pkg/types"
 	authdb "github.com/e2b-dev/infra/packages/db/pkg/auth"
 	"github.com/e2b-dev/infra/packages/shared/pkg/apierrors"
