+++ packages/dashboard-api/Dockerfile
@@ -30,7 +30,6 @@ RUN go mod download
 WORKDIR /build
 
 COPY ./auth/pkg ./auth/pkg
-COPY ./auth/internal ./auth/internal
 COPY ./clickhouse/pkg ./clickhouse/pkg
 COPY ./shared/pkg ./shared/pkg
 COPY ./db/client ./db/client
